requirements.txt
Flaskgunicornnumpy<2.0Pillowopencv-pythontflite-runtimegoogle-generativeai